+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # GoGoodwe - CHANGELOG
 
+## v2.0.6 (2024-06-7)
+
+- renamed SemsDataConstraint
 
 ## v2.0.5 (2024-05-10)
 
 - Optimised and refactored code to reduce CPU usage and memory pressure and increased speed.
-
-
 
 ## v2.0.3 (2024-04-26)
