+++ internal/pkg/types/dailysummarydata.go
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-package monitordata
+package types
 
 // DailySummaryData - Struct to hold daily summary data
 type DailySummaryData struct {
